==> 4.resultsByAgeGroup+VisualRepresentation.ipynb <==
## Sex differences in Autism Spectrum Disorder, a Comorbidity Pattern Analysis in National Scale Data:  (iv) age comparison, visual representation

As a result for our analysis, we will have a list of PheCodes that are more likely in ASD females in a specific age range. We analyzed 4 different age groups:
- 0 to 2 years old
- 3 to 5 years old
- 6 to 11 years old
- 12 to 18 years old

In this last step, we will look for those PheCodes specific for each age range and which one are statistically significant in more than one group. 

We will use UpSetR (http://caleydo.org/tools/upset/).


First, we install and load the UpSetR package. 

install.packages( "UpSetR" )
library( UpSetR )

Then, we load the results generated in 3.ComapreThreeGroupsResults

load( "from0to2.RData" )
load( "from3to5.RData" )
load( "from6to11.RData" )
load( "from12to18.RData" )

Finally, we format the input as a list to apply the "upset" function in the following way: 

listInput <- list( From0to18  = c( from0to18$description ), 
                   From0to2   = c( from0to2$description ),  
                   From3to5   = c( from3to5$description ), 
                   From6to11  = c( from6to11$description ), 
                   From12to18 = c( from12to18$description )
                   )

upset( fromList( listInput ), 
       order.by       = "freq", 
      mainbar.y.label = "Phenotype Intersections", 
      sets.x.label    = "Stat.Sign. Phenotype Per Age Group" )
